refactor: Smaller styles modules
